Create README.md
Practical 1. Python program to convert the temperature in degree centigrade to Fahrenheit.
Practical 2. Python program to find the area of a triangle whose sides are given.
Practical 3. Python program to find out the average of a set of integers.
Practical 4. Python program to find the product of a set of real numbers.
Practical 5. Python program to find the circumference and area of a circle with a given radius.
Practical 6. Python program to check whether the given integer is a multiple of both 5 and 7.
Practical 7. Python program to find the average of 10 numbers using while loop.
Practical 8. Python program to display the given integer in a reverse manner.
Practical 9. Python program to find the geometric mean of n numbers.
Practical 10. Python program to find the roots of a quadratic equation.